requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1
clickhouse-driver==0.2.6
pgvector==0.2.4

# Message Queue
nats-py==2.6.0
redis==5.0.1

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# Data Processing
pandas==2.1.4
numpy==1.26.2
scipy==1.11.4

# ML/AI
scikit-learn==1.3.2
sentence-transformers==2.2.2
openai==1.3.7
anthropic==0.7.8

# Monitoring & Logging
prometheus-client==0.19.0
structlog==23.2.0
opentelemetry-distro==0.43b0
opentelemetry-instrumentation-fastapi==0.43b0

# Configuration & Secrets
python-dotenv==1.0.0
cryptography==41.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
mypy==1.7.1
flake8==6.1.0

# Utilities
python-ulid==2.2.0
pydantic==2.5.0
tenacity==8.2.3
